cat blob: Don't fill blob buffers we don't use
To cat a single blob, copy the ones we don't want to io::sink() until we
get to the one we do. Check that the hashes match while we're at it.

We'll want something more sophisticated for actual tree walks/restore.
Maybe some cache directory of uncompressed packs, and an in-memory tree
cache (given that trees will be much smaller)?