Cache backends (with a CachedBackend of course!)
Remote backends (S3, B2, etc.) will want to use a handful of local files
as a write-through cache so that objects are seekable (let's do away
with all the SeekableReader trait nonsense) and, well, cached.
We don't want have to hit the server if we frequently access a handful
of objects (like indexes and the tree packs).

It would be pretty straightforward to download files to some
~/.cache/backpak/objects and evict the oldest ones once we exceed a
certain size... but that sounds like some work; let's do it another day.

For local (filesystem) backends, just read and write the files directly.
Easy, and avoids needless copies.