Move CLI files into ui module
The separation isn't as clean as it could be - ideally ui/ would be a
pretty thin layer that just handles CLI parsing, but ui/backup.rs is all
the crazy concurrency plumbing needed for running a backup.

Still seems to beat having snapshot.rs and snapshots.rs (the file
defining snapshots, and a command to list snapshots, respectively)
sitting together in the top-level src/ directory.