Refactor check, move chunks_in_tree() into tree.rs
Also add chunks_in_forest(), which we can use to gather all chunks
reachable by a snapshot. Combine that with the keys of the forest
and tada! All the reachable blobs in a snapshot.
Do that for each snapshot, and we have the set of live blobs.
We can GC the rest when we prune the repository.