Try giving backend interior mutability
So far, backends have had mutability that matches their semantics:

read() and list() are &self, write() and remove() are &mut self.
This becomes a problem when we're pruning: A BlobReader needs an
immutable reference to the cached backend to read from it,
but we'll also be mutating as we go, making new packs!

So, we need some sort of interior mutability. But where to put it?
We could wrap the cached backend in some Mutex/RWLock in `backup.rs`,
but that seems really noisy. And for things (like the FS backend)
whose mutability is outside this process, we'd be needlessly locking.
Instead, we could put the interior mutability in the backends
themselves...