tree: Don't compare metadata by atime in their Eq
It changes on the filesystem when you look at it, so using it to diff
trees (especially against the FS tree) produces... unhelpful input.

Each backup still captures a fresh copy of all the metadata (trees are
cheap and we've already called stat()) so we're not throwing any
data out with this change. Nor does it impact file_changed() and the
heuristics we use to see when we can recycle a past snapshot's chunks.