Async misadventures
An initial, crude switch to async code.
I resisted doing this for a while, out of fears that it would add a lot
of complexity, but in the end it's probably a better concurrency model
than manually managing threads, _even with_ tools like Rayon to
intelligently parallelize the CPU-heavy bits.

Some follow-up passes would almost certainly pay dividends - I'm not
sure if it'll be beneficial to make all the I/O asynchronous, but surely
there's some more places we can put it to good use.