Damn, Camino is cool
If we want to provide portable backups between Unix and Windows, we run
into the Makefile Problem - there's no portable way to encode file paths
_in your own data_.

Let's just mandate everything be UTF-8 with Camino. This simplifies a
_bunch_ of code and encodes to the exact same bytes as before!

The only thing we should double check is that we fail fast if something
we wanna back up contains non-UTF8 paths. Should be a simple test.