metadata: Don't record what we can't compare or restore
Change time is pretty meaningless to us in POSIX - since it's the
time the *metadata* changed, we can't set it, nor can we compare it
between snapshots.

Instead of the awkward situtation where we'd get different trees
when it changed, but ignore those differences when diffing/restoring,
just leave it off. Thanks to serde magic we won't even miss it from
existing test repos!